add next-web-server
