+++ packages/next/server/dev/hot-reloader.ts
@@ -543,6 +543,7 @@ export default class HotReloader {
                     canonicalBase: this.config.amp.canonicalBase,
                     i18n: this.config.i18n,
                     previewProps: this.previewProps,
+                    distDir: this.config.distDir,
                   } as any)}!`,
                   isServer: false,
                   isServerWeb: true,
