+++ test/integration/react-streaming-and-server-components/app/package.json
@@ -2,6 +2,7 @@
   "private": true,
   "scripts": {
     "lnext": "node -r ../../react-18/test/require-hook.js ../../../../packages/next/dist/bin/next",
+    "debug": "NODE_OPTIONS='--inspect' node -r ../../react-18/test/require-hook.js ../../../../packages/next/dist/bin/next",
     "dev": "yarn lnext dev",
     "build": "yarn lnext build",
     "start": "yarn lnext start"
